make relapseTime unit testable and test it
